server/requirements.txt
-i https://pypi.org/simple
alembic==1.11.1
aniso8601==9.0.1
asttokens==2.2.1
attrs==23.1.0 ; python_version >= '3.7'
backcall==0.2.0
bcrypt==4.0.1 ; python_version >= '3.6'
beautifulsoup4==4.12.2
blinker==1.6.2 ; python_version >= '3.7'
certifi==2023.5.7 ; python_version >= '3.6'
charset-normalizer==3.2.0 ; python_full_version >= '3.7.0'
click==8.1.5 ; python_version >= '3.7'
decorator==5.1.1 ; python_version < '3.11' and python_version >= '3.7'
exceptiongroup==1.1.2 ; python_version < '3.11'
executing==1.2.0
flask==2.3.2
flask-bcrypt==1.0.1
flask-cors==4.0.0
flask-login==0.6.2
flask-migrate==4.0.4
flask-restful==0.3.10
flask-sqlalchemy==3.0.5
greenlet==2.0.2 ; platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
gunicorn==21.0.1
h11==0.14.0 ; python_version >= '3.7'
honcho==1.1.0
idna==3.4 ; python_version >= '3.5'
importlib-metadata==6.8.0 ; python_version < '3.9'
importlib-resources==6.0.0
ipdb==0.13.13
ipython==8.12.2 ; python_version < '3.11' and python_version >= '3.7'
itsdangerous==2.1.2 ; python_version >= '3.7'
jedi==0.18.2 ; python_version >= '3.6'
jinja2==3.1.2 ; python_version >= '3.7'
mako==1.2.4 ; python_version >= '3.7'
markupsafe==2.1.3 ; python_version >= '3.7'
matplotlib-inline==0.1.6 ; python_version >= '3.5'
outcome==1.2.0 ; python_version >= '3.7'
packaging==23.1 ; python_version >= '3.7'
parso==0.8.3 ; python_version >= '3.6'
pexpect==4.8.0 ; sys_platform != 'win32'
pickleshare==0.7.5
prompt-toolkit==3.0.39 ; python_full_version >= '3.7.0'
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pygments==2.15.1 ; python_version >= '3.7'
pysocks==1.7.1
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
selenium==4.10.0
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.3.0 ; python_version >= '3.7'
sortedcontainers==2.4.0
soupsieve==2.4.1 ; python_version >= '3.7'
sqlalchemy==2.0.19
sqlalchemy-serializer==1.4.1
stack-data==0.6.2
tomli==2.0.1 ; python_version < '3.11' and python_version >= '3.7'
traitlets==5.9.0 ; python_version >= '3.7'
trio==0.22.2 ; python_version >= '3.7'
trio-websocket==0.10.3 ; python_version >= '3.7'
typing-extensions==4.7.1 ; python_version >= '3.7'
urllib3==2.0.3 ; python_version >= '3.7'
wcwidth==0.2.6
werkzeug==2.3.6 ; python_version >= '3.8'
wsproto==1.2.0 ; python_full_version >= '3.7.0'
zipp==3.16.2 ; python_version < '3.10'
